Moved base64 encoding extension. Reformatted a few todos
